backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
psycopg[binary]>=3.1.0
marshmallow==3.20.1
python-dotenv==1.0.0
Pillow>=10.2.0
bcrypt>=4.1.2
